+++ src/utils/loon.ts
@@ -20,6 +20,7 @@ const {
   trojanFilter,
   vmessFilter,
   wireguardFilter,
+  vlessFilter,
 } = internalFilters
 const logger = createLogger({ service: 'surgio:utils:loon' })
 
@@ -93,22 +94,29 @@ export const getLoonNodes = function (
           return config.join(',')
         }
 
+        case NodeTypeEnum.Vless:
         case NodeTypeEnum.Vmess: {
           if (
             !LOON_SUPPORTED_VMESS_NETWORK.includes(nodeConfig.network as any)
           ) {
             logger.warn(
-              `Loon 不支持 ${nodeConfig.network} 的 Vmess 节点，节点 ${nodeConfig.nodeName} 会被省略`,
+              `Loon 不支持 ${
+                nodeConfig.network
+              } 的 ${nodeConfig.type.toUpperCase()} 节点，节点 ${
+                nodeConfig.nodeName
+              } 会被省略`,
             )
             return void 0
           }
 
           const config: Array<string | number> = [
-            `${nodeConfig.nodeName} = vmess`,
+            `${nodeConfig.nodeName} = ${
+              nodeConfig.type === NodeTypeEnum.Vmess ? 'vmess' : 'VLESS'
+            }`,
             nodeConfig.hostname,
             nodeConfig.port,
             nodeConfig.method === 'auto'
-              ? `method=chacha20-ietf-poly1305`
+              ? `method=chacha20-poly1305`
               : `method=${nodeConfig.method}`,
             JSON.stringify(nodeConfig.uuid),
             `transport=${nodeConfig.network}`,
@@ -141,7 +149,10 @@ export const getLoonNodes = function (
             }
           }
 
-          if (nodeConfig.tls) {
+          if (
+            (nodeConfig.type === NodeTypeEnum.Vmess && nodeConfig.tls) ||
+            nodeConfig.type === NodeTypeEnum.Vless
+          ) {
             config.push(`over-tls=true`)
 
             if (nodeConfig.sni) {
@@ -306,7 +317,8 @@ export const getLoonNodeNames = function (
         httpFilter(item) ||
         httpsFilter(item) ||
         trojanFilter(item) ||
-        wireguardFilter(item),
+        wireguardFilter(item) ||
+        vlessFilter(item),
     ),
     filter,
   )